Remove config saving, not needed after all
It was designed to store OAuth access key and secret after fetching them
based on a consumer key and secret. Unfortunately, that does not work
for our case since we cannot supply consumer key and secret with the
software (they would be readable in plain text). We must limit ourselves
to using single access token, in which the user configures all the keys
herself.